feat(user): 更新用户头像功能
- 修改 UserFileController 中的 updateImage 方法，返回 UserVo 对象
- 新增 finishUpdate 方法处理头像更新逻辑
- 更新 LoginUserMapper 接口，将参数类型从 Integer 改为 Long
- 在 LoginUserMapper.xml 中添加头像文件 ID 相关的 SQL 语句
- 更新 LoginServiceImpl 中的 getPermission 方法，参数类型改为 Long

+++ spring-server/api/api-model/src/main/java/com/czy/api/api/user_relationship/LoginService.java
@@ -42,5 +42,5 @@ public interface LoginService {
     LoginUserRequest findBackUserPassword(String account, String password);
 
     // permission查询
-    Integer getPermission(Integer userId);
+    Integer getPermission(Long userId);
 }
